02_langchain_tools_agent/requirements.txt
langchain==0.3.12
langchain-community==0.3.12
langchain-snowflake==0.1.3
snowflake-ml-python==1.7.4
snowflake-snowpark-python==1.28.0
streamlit==1.39.0